Extract CLI argument parser
